refactor: 1. 대표님 피드백 받은 것 리펙터링
