+++ ggtics-recruitement-lol-score/src/main/kotlin/com/your/lol/domain/Stat.kt
@@ -1,0 +1,7 @@
+package com.your.lol.domain
+
+class Stat(
+        val winRate: Double,
+        val championDataDtos: List<ChampionDataDto>) {
+    private constructor() : this(0.0, emptyList())
+}
