+++ ggtics-recruitement-lol-score/src/main/kotlin/com/your/lol/domain/dto/PreProcessedData.kt
@@ -1,0 +1,17 @@
+package com.your.lol.domain.dto
+
+/**
+ * 도메인 내부에서 사용되고 있는 데이터 가공목적으로 사용한 DTO 클래스
+ */
+data class PreProcessedData(
+        val championName: String,
+        val teamId: Int,
+        val kda: Double,
+        val deathsByEnemyChamps: Long,
+        val gameDuration: Long,
+        val teamWin: Boolean) {
+
+    fun convertTeamWinToNumber(): Int {
+        return if (teamWin) 1 else 0
+    }
+}
