+++ ggtics-recruitement-lol-score/src/main/kotlin/com/your/lol/dto/riot/ChallengesDto.kt
