+++ ggtics-recruitement-lol-score/src/main/resources/application.yml
