--- lol/src/main/kotlin/com/your/lol/presentation/SummonerStatsController.kt
@@ -1,23 +1,0 @@
-package com.your.lol.presentation
-
-import com.your.lol.domain.LoLScoreStatsCalculator
-import com.your.lol.dto.statistics.StatsDto
-import lombok.RequiredArgsConstructor
-import org.springframework.http.HttpStatus
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.ResponseStatus
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-@RequestMapping("/api/")
-@RequiredArgsConstructor
-class SummonerStatsController(
-        private val loLScoreStatsCalculator: LoLScoreStatsCalculator) {
-
-    @RequestMapping("/summoner/{summonerName}")
-    @ResponseStatus(HttpStatus.OK)
-    fun findStatisticsBySummonerName(@PathVariable summonerName: String): StatsDto {
-        return loLScoreStatsCalculator!!.calculateStatsBySummonerName(summonerName)
-    }
-}
